+++ ansibleplaybookgrapher/version.py
@@ -1,2 +1,2 @@
-__version__ = "2.2.0-dev"
+__version__ = "2.2.0.dev0"
 __prog__ = "ansible-playbook-grapher"
