+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-description-file = Readme.md
+description_file = Readme.md
 
 [bdist_wheel]
 universal = 1
